Simplify target associations
